+++ src/Clinica Frba/Clases/Usuarios.cs
@@ -25,7 +25,7 @@ namespace Clinica_Frba.Clases
                 {
                     Rol unRol = new Rol();
                     unRol.Nombre = (string)lector["nombre"];
-                    unRol.Id = (decimal)lector["rol"];
+                    unRol.Id = (int)lector["rol"];
                     
                     unRol.Habilitado = true;
                     Lista.Add(unRol);
